- Showing sale price if present alongside original price - Styling for sale price and original price - New Price component for better organization
